pyproject.toml
[project]
name = "just_facebook_mcp"
version = "0.0.0"
description = "MCP server for automating and managing interactions on a Facebook Page using the Facebook Graph API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Livia Zaharia", email = "liviazaharia2020@gmail.com" },
    { name = "HagaiHen"}
]
keywords = ["mcp", "facebook", "api", "social-media", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp",
    "python-dotenv",
    "requests",
]



[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]
build = [
    "build",
    "twine",
    "hatch",
]



[project.scripts]
just_facebook_mcp = "just_facebook_mcp:main"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"




[tool.hatch.version]
path = "just_facebook_mcp/__init__.py"
pattern = '__version__ = "(?P<version>[^"]+)"'

[tool.hatch.build.targets.wheel]
packages = ["just_facebook_mcp"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 




[project.urls]
Homepage = "https://github.com/Livia-Zaharia/just_facebook_mcp"
Repository = "https://github.com/Livia-Zaharia/just_facebook_mcp"
Issues = "https://github.com/Livia-Zaharia/just_facebook_mcp/issues"

